requirements.txt
# Core
python-dotenv~=1.0.1
aiogram~=3.22.0
SQLAlchemy~=2.0.43
alembic~=1.16.4
aiohttp~=3.12.14
redis~=6.4.0

# Database drivers
asyncpg~=0.30.0
aiosqlite~=0.20.0

# Admin panel
sqladmin>=0.20
starlette
uvicorn
itsdangerous

# Validation & Security
pydantic~=2.5.0
cryptography~=41.0.7

# Utils
python-dateutil~=2.8.2

# Testing
pytest~=8.4.2
pytest-asyncio~=0.24.0
pytest-cov~=4.1.0
pytest-mock~=3.12.0
